Move sparse/ out of defs/
