+++ src/PAOFLOW/sparse/adaptive_smearing.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 import numpy as np
 
 if TYPE_CHECKING:
-    from ...DataController import DataController
+    from ..DataController import DataController
 
 
 def do_adaptive_smearing(
