+++ src/PAOFLOW/sparse/fermi.py
@@ -8,7 +8,7 @@ from scipy import sparse
 from scipy.sparse import linalg as spla
 
 if TYPE_CHECKING:
-    from ...DataController import DataController
+    from ..DataController import DataController
 
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
